+++ src/RedSound/RedMidiCtrl.cpp
@@ -1152,7 +1152,7 @@ static void __MidiCtrl_Wave(RedSoundCONTROL*, RedKeyOnDATA*, RedTrackDATA* track
     track->m_waveData = 0;
     track->m_waveBase = 0;
     waveNo = *track->m_command++;
-    if (((u32)track->m_waveBankData != 0) && (waveNo < track->m_waveBankData->m_tableCount)) {
+    if ((track->m_waveBankData != 0) && (waveNo < track->m_waveBankData->m_tableCount)) {
         waveTable = (int)track->m_waveBankData->m_waveOffsets;
         track->m_waveData = (RedWaveDATA*)((int)track->m_waveBankData + *(int*)(waveTable + waveNo * 4));
         track->m_waveBase = track->m_waveBankData->m_aramAddress;
